Better broadcastVariable
I couldn't figure out a way to define this in base.Qoid or Index using an interface...but that's okay, for now this is fine. It's only a little bit of repeated code.

I went all the way down to Properties so that I can implement stuff like a custom variable symbol different than @, while modifying the least code possible.